+++ jlrs/tests/abstract_fields.rs
@@ -4,6 +4,7 @@ mod util;
 mod tests {
     use super::util::JULIA;
     use jlrs::prelude::*;
+    use jlrs::wrappers::ptr::ValueRef;
 
     #[test]
     fn read_abstract_field() {
@@ -24,7 +25,7 @@ mod tests {
                     .instantiate(&mut *frame, &mut [arg1])?
                     .into_jlrs_result()?;
 
-                let a = instance.get_raw_field::<Value, _>("a")?;
+                let a = unsafe { instance.get_raw_field::<ValueRef, _>("a")?.value().unwrap() };
                 assert_eq!(a.unbox::<u32>()?, 3);
 
                 Ok(())