+++ CHANGELOG.md
@@ -32,3 +32,6 @@
  - The `AsUnrooted` trait has been removed.
  
  - Most of the extensions defined in the extensions module have moved: `jlrs::extensions::f16` to `jlrs::wrappers::inline::f16`. `jlrs::extensions::ndarray` to `jlrs::ndarray`, `jlrs::extensions::multitask` to `jlrs::multitask`, and `jlrs::extensions::pyplot` to `jlrs::pyplot`.
+
+ - Pointer wrapper types don't implement `ValidLayout`, only `Ref` and inline wrappers do.
+